Address Samir's comments
